todo/mysite/requirements.txt
coverage==4.4.2
flake8==3.5.0
Django==2.0.1
django-extensions==1.9.9
django-filter==1.1.0
djangorestframework==3.7.7
django-cors-headers==2.1.0
Markdown==2.6.11
mysqlclient==1.3.12
pytz==2017.3
PyJWT==1.5.3
requests==2.18.4
zappa==0.45.1
django-mysql==2.2.2
itsdangerous==0.24
